Refactor UI and Interactables Handling
This commit introduces a series of enhancements and refactoring across the UI and interactables handling within the project. Key changes include:

- Enhanced the `Button` default constructor in `Button_PG.h` to support default initialization with an empty lambda function, improving ease of use.
- Updated `InGameHUD_PG.h` with significant changes: removed and renamed methods for clarity, added direct handling of `settingsButton` and `TextDisplayBox` for dynamic HUD updates, and introduced new methods for interaction and updating ship properties.
- Redesigned `Menu_PG.h` towards an abstract base class structure, enforcing a consistent interface for menus. This includes specialized methods for adding different types of interactables and handling menu item interactions.
- Improved `OptionsMenu_PG.h` with direct methods for adding interactables and made the `update` method concrete, reflecting a more modular approach to options menu construction.
- Enhanced `StartMenu_PG.h` with new interaction and drawing methods.
- In `WaterTiler_PG.h`, improved tile positioning precision by explicitly casting positions to `float`.
- Adjusted enemy ship destination offset in `EnemyShipHandler_PG.cpp` for better positioning.
- Implemented dynamic menu updating in `MenuHandler_PG.cpp` based on `menuType`.
- Across multiple files, general refactoring and cleanup were performed to remove outdated methods and replace them with more flexible implementations, moving towards a more object-oriented and modular design.
- Specific updates in `InGameHUD_PG.h` include changes to `addInteractablesToMenu` and `updateShipPropertiesString` methods for direct object management and simplified HUD updates.

+++ PirateGame/Header/WaterTiler_PG.h
@@ -55,7 +55,7 @@ namespace PirateGame {
 
             for (int x = startX; x <= endX; ++x) {
                 for (int y = startY; y <= endY; ++y) {
-                    waterTileSprite.setPosition(x * scaledFrameWidth, y * scaledFrameHeight);
+                    waterTileSprite.setPosition(static_cast<float>(x * scaledFrameWidth), static_cast<float>(y * scaledFrameHeight));
                     window->draw(waterTileSprite);
                 }
             }
